+++ CHANGELOG.rst
@@ -30,6 +30,7 @@ Changed
 - narwhalified BaseNumericTransformer `#358 <https://github.com/lvgig/tubular/issues/358>_`
 - narwhalified BaseCappingTransformer `#357 <https://github.com/lvgig/tubular/issues/357>_`
 - narwhalified CappingTransformer `#361 <https://github.com/lvgig/tubular/issues/361>_`
+- narwhalified OutOfRangeNullTransformer `#362 <https://github.com/lvgig/tubular/issues/362>_`
 - narwhalified MeanImputer `#344 https://github.com/lvgig/tubular/issues/344_`
 - narwhalified BaseNumericTransformer `#358 https://github.com/lvgig/tubular/issues/358`
 - narwhalified DropOriginalMixin `#352 <https://github.com/lvgig/tubular/issues/352>_`